update "added components for sales (seller)"
